- fixed some code hygiene
