+++ CompositeState/CompositeState/Table/StateTransitionTable.cs
@@ -5,23 +5,6 @@ using System.Linq;
 namespace CompositeState.Table
 {
 
-    [System.Diagnostics.DebuggerDisplay("{DebuggerDisplay}")]
-    public class StateTuple
-    {
-        public Enum[] State { get; set; }
-        public TransitionTuple[] Transitions { get; set; }
-        public string DebuggerDisplay { get; set; }
-    }
-
-    [System.Diagnostics.DebuggerDisplay("{DebuggerDisplay}")]
-    public class TransitionTuple
-    {
-        public Enum Input { get; set; }
-        public int Next { get; set; }
-        public Action Output { get; set; }
-        public string DebuggerDisplay { get; set; }
-    }
-
     public class StateTransitionTable : IStateMachine
     {
         private int currentState = 0;
