+++ CompositeState/CompositeState/Table/StateTuple.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace CompositeState.Table
+{
+
+    [System.Diagnostics.DebuggerDisplay("{DebuggerDisplay}")]
+    public class StateTuple
+    {
+        public Enum[] State { get; set; }
+        public TransitionTuple[] Transitions { get; set; }
+        public string DebuggerDisplay { get; set; }
+    }
+
+}
