+++ CompositeState/CompositeState/Table/TransitionTuple.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace CompositeState.Table
+{
+
+    [System.Diagnostics.DebuggerDisplay("{DebuggerDisplay}")]
+    public class TransitionTuple
+    {
+        public Enum Input { get; set; }
+        public int Next { get; set; }
+        public Action Output { get; set; }
+        public string DebuggerDisplay { get; set; }
+    }
+
+}
